Add files via upload
Example - Json schema for EPG - filled in
Configuration
The details for a feed to the schedule needs to be provided. The integration is configured with a list of URLs from which to fetch EPG content. The URLs should serve a json file that follows a schema similar to the examples below in order for this integration to properly parse the EPG data.